refactor: move webhook/escalation into escalation
